+++ src/index.js
@@ -29,6 +29,7 @@ import experimentation from './experimentation/handler.js';
 import conversion from './conversion/handler.js';
 import essExperimentationDaily from './experimentation-ess/daily.js';
 import essExperimentationAll from './experimentation-ess/all.js';
+import costs from './costs/handler.js';
 
 const HANDLERS = {
   apex,
@@ -43,6 +44,7 @@ const HANDLERS = {
   conversion,
   'experimentation-ess-daily': essExperimentationDaily,
   'experimentation-ess-all': essExperimentationAll,
+  costs,
 };
 
 function getElapsedSeconds(startTime) {